Add files via upload
So I can't export my character rig as-is in the FBX file format, because it doesn't support some constraints and transforms. I don't believe that will be a problem when importing the file into Unity, because all of the joint information will be baked in anyway. I don't want to mess with this too much at this unfinished state, but I will put something online that may be of use.